Split WalletSync by Activaton & Manage side content
