+++ apps/ledger-live-desktop/src/renderer/screens/settings/sections/General/WalletSync/SideContent/Manage.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const WalletSyncManage = () => {
+  return <div>Wallet Sync</div>;
+};
+
+export default WalletSyncManage;
